requirements.txt
python-telegram-bot
pydantic
pyyaml
